Contianer : Developer features.
IsLocked( with password ), IsWord( is set as world container ),
IsGenerated( is genarated by prop or shits )

Also, empty contianer will be removed when you loaded the map. ( World
Container is can be saved whether it has item or not. )

World Container = Useful for container specific features like generating
random items in the container.

Generated Container = It's useful to prop -> contianer. if it's
generated, the container will not be saved.