Make character var registration create table fields
This means we can add more fields arbitrarily without having to migrate the db manually. It also allows schemas to register their own vars that are persistent, and not dependent on shoving more entries in the JSON data field.

Add InsertIgnore and Alter to mysql library
Change names of some bool variables to follow bName convention