Don't remove items with invalid data from database
It's better to have stale items in the database and simply not load them as part of an inventory, rather than deleting them. This avoids destructive operations - useful for when an inventory is accidentally resized and out-of-bounds items are removed and cannot be restored after the inventory has been reverted back to its original size.