Fixed previous character ban preventing later checks from happening (#442)
* Fixed previous character ban preventing later checks from happening

In the CanPlayerUseCharacter hook, there is a check to make sure the character isnt banned. If the character received a temporary ban at some point and that ban runs out, the check in that function, returns, instead of letting later checks from happening, which could for instance let a character without a whitelist still load into a whitelisted character.