Fix pacAdjust data not being removed correctly
In cases where pacAdjust returned results significantly different from the original pacData, RemovePart() would not call RemovePACPart() on the correct table and thus the part would become permanently stuck to the player. This ensures that the pacData table passed to RemovePACPart matches the one initially attached by PAC, if needed.