Run InventoryItemRemoved hook in ITEM:Remove()
InventoryItemRemoved was previously called only when INVENTORY:Remove() was used and not when the item was removed directly. Both already called ITEM:OnRemoved(), but there's no way to execute a hook on items removed via the traditional method despite the support function for it existing.

This simply runs the InventoryItemRemoved hook inside of ITEM:Remove() when bNoDelete is false. An alternative solution would be the creation of an OnItemDeleted hook, similar to OnItemSpawned (even though that acts directly on the created entity and not the instance removed).